Adding Scrutinizer
Scrutinizer is an automated code quality tool, you can see it's features at https://scrutinizer-ci.com/tour/measure-and-improve-code-quality

In this pull I'm adding the scrutinizer image that takes to the quality report. Still haven't finished the first check, we will need to wait a bit.

Let's see if it provides interesting information.